+++ packages/node/src/build-info.ts
@@ -3,11 +3,15 @@ import type { PandaContext } from './create-context'
 
 export async function buildInfo(ctx: PandaContext, outfile: string) {
   const { filesWithCss, files } = ctx.parseFiles()
-
   logger.info('cli', `Found ${colors.bold(`${filesWithCss.length}/${files.length}`)} files using Panda`)
 
-  const minify = ctx.config.minify
+  const { minify, staticCss } = ctx.config
   logger.info('cli', `Writing ${minify ? '[min] ' : ' '}${colors.bold(outfile)}`)
+
+  if (staticCss) {
+    logger.info('cli', 'Adding staticCss definitions...')
+    ctx.staticCss.process(staticCss)
+  }
 
   const output = JSON.stringify(ctx.encoder.toJSON(), null, minify ? 0 : 2)