+++ packages/types/src/theme.ts
@@ -1,4 +1,4 @@
-import type { LayerStyles, TextStyles } from './composition'
+import type { AnimationStyles, LayerStyles, TextStyles } from './composition'
 import type { RecipeConfig, SlotRecipeConfig } from './recipe'
 import type { CssKeyframes } from './system-types'
 import type { SemanticTokens, Tokens } from './tokens'
@@ -28,6 +28,10 @@ export interface Theme {
    * The layer styles for your project.
    */
   layerStyles?: LayerStyles
+  /**
+   * The animation styles for your project.
+   */
+  animationStyles?: AnimationStyles
   /**
    * Multi-variant style definitions for your project.
    * Useful for defining component styles.