+++ packages/logger/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @pandacss/logger
 
+## 0.49.0
+
+### Patch Changes
+
+- Updated dependencies [97a0e4d]
+  - @pandacss/types@0.49.0
+
+## 0.48.1
+
+### Patch Changes
+
+- @pandacss/types@0.48.1
+
+## 0.48.0
+
+### Patch Changes
+
+- @pandacss/types@0.48.0
+
 ## 0.47.1
 
 ### Patch Changes