+++ packages/node/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @pandacss/node
 
+## 0.49.0
+
+### Patch Changes
+
+- Updated dependencies [97a0e4d]
+  - @pandacss/generator@0.49.0
+  - @pandacss/types@0.49.0
+  - @pandacss/core@0.49.0
+  - @pandacss/config@0.49.0
+  - @pandacss/parser@0.49.0
+  - @pandacss/logger@0.49.0
+  - @pandacss/token-dictionary@0.49.0
+  - @pandacss/extractor@0.49.0
+  - @pandacss/shared@0.49.0
+
+## 0.48.1
+
+### Patch Changes
+
+- fd87f3a: Fix issue where `staticCss` artifacts were not included in the build info json.
+- Updated dependencies [af9715a]
+  - @pandacss/generator@0.48.1
+  - @pandacss/config@0.48.1
+  - @pandacss/parser@0.48.1
+  - @pandacss/core@0.48.1
+  - @pandacss/extractor@0.48.1
+  - @pandacss/logger@0.48.1
+  - @pandacss/shared@0.48.1
+  - @pandacss/token-dictionary@0.48.1
+  - @pandacss/types@0.48.1
+
+## 0.48.0
+
+### Patch Changes
+
+- Updated dependencies [2bc12d2]
+  - @pandacss/generator@0.48.0
+  - @pandacss/config@0.48.0
+  - @pandacss/parser@0.48.0
+  - @pandacss/core@0.48.0
+  - @pandacss/extractor@0.48.0
+  - @pandacss/logger@0.48.0
+  - @pandacss/shared@0.48.0
+  - @pandacss/token-dictionary@0.48.0
+  - @pandacss/types@0.48.0
+
 ## 0.47.1
 
 ### Patch Changes