+++ packages/token-dictionary/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @pandacss/token-dictionary
 
+## 0.49.0
+
+### Patch Changes
+
+- Updated dependencies [97a0e4d]
+  - @pandacss/types@0.49.0
+  - @pandacss/logger@0.49.0
+  - @pandacss/shared@0.49.0
+
+## 0.48.1
+
+### Patch Changes
+
+- @pandacss/logger@0.48.1
+- @pandacss/shared@0.48.1
+- @pandacss/types@0.48.1
+
+## 0.48.0
+
+### Patch Changes
+
+- @pandacss/logger@0.48.0
+- @pandacss/shared@0.48.0
+- @pandacss/types@0.48.0
+
 ## 0.47.1
 
 ### Patch Changes